split up main.js
